[bbr] rename file name with bbr_ prefix (#5240)
