+++ BUILD.gn
@@ -66,8 +66,8 @@ static_library("lib-ot-core") {
     "src/core/api/thread_api.cpp",
     "src/core/api/thread_ftd_api.cpp",
     "src/core/api/udp_api.cpp",
-    "src/core/backbone_router/leader.cpp",
-    "src/core/backbone_router/local.cpp",
+    "src/core/backbone_router/bbr_leader.cpp",
+    "src/core/backbone_router/bbr_local.cpp",
     "src/core/coap/coap.cpp",
     "src/core/coap/coap_message.cpp",
     "src/core/coap/coap_secure.cpp",
