+++ src/core/CMakeLists.txt
@@ -102,8 +102,8 @@ set(COMMON_SOURCES
     api/thread_api.cpp
     api/thread_ftd_api.cpp
     api/udp_api.cpp
-    backbone_router/leader.cpp
-    backbone_router/local.cpp
+    backbone_router/bbr_leader.cpp
+    backbone_router/bbr_local.cpp
     coap/coap.cpp
     coap/coap_message.cpp
     coap/coap_secure.cpp
